% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_centroid.R
\name{centroid}
\alias{centroid}
\alias{centroid.numeric}
\alias{centroid.matrix}
\alias{centroid.data.frame}
\title{Find centroid}
\usage{
centroid(x, ...)

\method{centroid}{numeric}(x, ...)

\method{centroid}{matrix}(x, ...)

\method{centroid}{data.frame}(x, ...)
}
\arguments{
\item{x}{matrix or data frame with observations in rows and dimensions in columns}

\item{...}{for data frame method, columns that store cooridinates}
}
\value{
numeric vector of centroid coordinates
}
\description{
Generic function that, given a set of pionts, finds their centroid.
}
\details{
Centroid coordinates are derived as average coordinates of all points in respective dimensions.
If a point is missing a coordinate, the remainig dimensions are still considered.
Methods exist for matrix and data frame.

The data frame method isolates relevant columns as a matrix and calls the matrix method.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: simply returns the mean, NAs are omitted

\item \code{matrix}: simply averages columns, omitting NAs

\item \code{data.frame}: extracts the requested columns and converts them to a matrix and then calls the matrix method
}}

\section{Column specification}{

Matrices are accepted as is, all columns are considered.
If \code{x} is a data frame, columns that carry coordinates may be specified.
They can be given as strings, bare names or numerical indices but
they must all be specified in the same way.
Also, each column must be specified separately, vectors are not accepted.
If no columns are specified, all numeric ones are considered.
}

