% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentroidR-package.R
\docType{package}
\name{CentroidR-package}
\alias{CentroidR}
\alias{CentroidR-package}
\title{CentroidR: CentroidR}
\description{
CentroidR provides the infrastructure to centroid profile spectra
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adafede/CentroidR}
  \item \url{https://adafede.github.io/CentroidR}
  \item Report bugs at \url{https://github.com/adafede/CentroidR/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{adafede@gmail.com} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

Authors:
\itemize{
  \item Johannes Rainer \email{Johannes.Rainer@eurac.edu} (\href{https://orcid.org/0000-0002-6977-7147}{ORCID})
}

}
\keyword{internal}
