% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_centroidr.R
\docType{package}
\name{centroidr}
\alias{centroidr}
\alias{centroidr-package}
\title{centroidr: centroids and distances}
\description{
The package finds the centroid of a set of points in n-dimensional space
and calculates the distances between the points and the centroid.
It provides two functinos: \code{centroid} and \code{cendist}, and has one internal function, \code{euclidean}.
}
\section{functions}{

\code{euclidean} is an internal function that calculte the euclidean distance between points.

\code{centroid} finds the centroid and returns its coordinates.

\code{centroid_distance} calls \code{centroid} and \code{euclidean}
and returns the distances between each point and the centroid.
}

\author{
Aleksander Chlebowski
}
