% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_euclidean.R
\name{euclidean}
\alias{euclidean}
\alias{distance}
\title{Find euclidean distance between two points}
\usage{
euclidean(pt1, pt2)
}
\arguments{
\item{pt1, pt2}{numeric vectors of length n; coordinates of points 1 and 2, respectively}
}
\value{
a single number
}
\description{
Given two sets of coordinates, find euclidean distance between them in n-dimensional space.
}
\details{
This is a simple vectorised calculation that returns the square root
of the sum of squared differences between respective elements of x and y.
It is used internally to calculate distances between points in a set and the set's centroid.
}
\keyword{internal}
