% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{setup_logger}
\alias{setup_logger}
\title{Setup Logging for Centroiding Process}
\usage{
setup_logger(dir = Sys.getenv("HOME"), filename = "centroiding.log")
}
\arguments{
\item{dir}{Directory. Defaults to HOME}

\item{filename}{Filename. Defaults to centroiding.log}
}
\description{
Setup Logging for Centroiding Process
}
\keyword{internal}
