% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{try_centroid_one_file}
\alias{try_centroid_one_file}
\title{Wrapper for centroiding with error handling}
\usage{
try_centroid_one_file(file, pattern, replacement, ...)
}
\arguments{
\item{file}{Path to the input mzML file.}

\item{pattern}{Pattern for modifying the file path.}

\item{replacement}{Replacement string for output path.}

\item{...}{Additional parameters to pass to \code{centroid_one_file}.}
}
\value{
Logical indicating success (\code{TRUE}) or failure (\code{FALSE}).
}
\description{
Wrapper for centroiding with error handling
}
\keyword{internal}
