% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{centroid_one_file}
\alias{centroid_one_file}
\title{Centroid an mzML file with configurable parameters}
\usage{
centroid_one_file(
  file,
  pattern,
  replacement,
  min_datapoints_ms1 = 5L,
  min_datapoints_ms2 = 1L,
  mz_tol_da_ms1 = 0.0025,
  mz_tol_da_ms2 = 0.0025,
  mz_tol_ppm_ms1 = 5,
  mz_tol_ppm_ms2 = 5,
  mz_fun_ms1 = base::mean,
  mz_fun_ms2 = base::mean,
  int_fun_ms1 = base::max,
  int_fun_ms2 = base::max,
  mz_weighted = TRUE,
  time_domain = TRUE,
  intensity_exponent = 3
)
}
\arguments{
\item{file}{character(1). Path to the input mzML file. Must be a valid, accessible mzML format file.}

\item{pattern}{character(1). Regular expression pattern to match in the input file path, used for modifying the output file path.}

\item{replacement}{character(1). Replacement string for altering the output file path based on the \code{pattern}.}

\item{min_datapoints_ms1}{integer(1). Minimum datapoints to be considered for MS1 data. Default: 5L.}

\item{min_datapoints_ms2}{integer(1). Minimum datapoints to be considered for MS2 data. Default: 1L.}

\item{mz_tol_da_ms1}{numeric(1). m/z tolerance in Daltons for MS1 data. Default: 0.0025.}

\item{mz_tol_da_ms2}{numeric(1). m/z tolerance in Daltons for MS2 data. Default: 0.0025.}

\item{mz_tol_ppm_ms1}{numeric(1). m/z tolerance in parts per million (ppm) for MS1. Default: 5.}

\item{mz_tol_ppm_ms2}{numeric(1). m/z tolerance in parts per million (ppm) for MS2. Default: 5.}

\item{mz_fun_ms1}{function. Function to aggregate m/z values within each peak group (MS1). Ignored if \code{mz_weighted = TRUE}. Default: base::mean.}

\item{mz_fun_ms2}{function. Function to aggregate m/z values within each peak group (MS2). Ignored if \code{mz_weighted = TRUE}. Default: base::mean.}

\item{int_fun_ms1}{function. Function to aggregate peak intensities within each peak group (MS1). Default: base::max.}

\item{int_fun_ms2}{function. Function to aggregate peak intensities within each peak group (MS2). Default: base::max.}

\item{mz_weighted}{logical(1). If \code{TRUE}, uses intensity-weighted mean for m/z value aggregation. Default: TRUE.}

\item{time_domain}{logical(1). If \code{TRUE}, uses square root for m/z weighting (TRUE for TOF). Default: TRUE.}

\item{intensity_exponent}{numeric(1). If \code{mz_weighted=TRUE}, the exponent of the intensity for m/z weighting. Default: 3.}
}
\value{
logical(1). Returns TRUE if centroiding was successful, otherwise FALSE.
}
\description{
This function processes an mzML file to apply centroiding with detailed controls
for peak picking, smoothing, and noise estimation. It allows fine-tuning of MS1 and MS2
peak detection, optimizing spectral data analysis for various experimental needs.
}
\details{
The function processes both MS1 and MS2 data with user-defined smoothing, peak-picking, and noise estimation.
File path modifications are supported via \code{pattern} and \code{replacement}.
}
\author{
Johannes Rainer, Adriano Rutz
}
