% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{centroid_one_file}
\alias{centroid_one_file}
\title{Centroid an mzML file with configurable parameters}
\usage{
centroid_one_file(
  file,
  pattern,
  replacement,
  min_datapoints_ms1 = 5L,
  min_datapoints_ms2 = 2L,
  mz_tol_da_ms1 = 0.002,
  mz_tol_da_ms2 = 0.005,
  mz_tol_ppm_ms1 = 5,
  mz_tol_ppm_ms2 = 10,
  mz_fun = base::mean,
  int_fun = base::max,
  mz_weighted = TRUE,
  time_domain = TRUE,
  intensity_exponent = 3
)
}
\arguments{
\item{file}{\code{character(1)}
Path to the input mzML file. Must be a valid, accessible mzML format file.}

\item{pattern}{\code{character(1)}
Regular expression pattern to match in the input file path, used for modifying the output file path.}

\item{replacement}{\code{character(1)}
Replacement string for altering the output file path based on the \code{pattern}.}

\item{min_datapoints_ms1}{\code{integer(1)} (default: \code{5L})
Minimum datapoints to be considered for MS1 data.}

\item{min_datapoints_ms2}{\code{integer(1)} (default: \code{2L})
Minimum datapoints to be considered for MS2 data.}

\item{mz_tol_da_ms1}{\code{numeric(1)} (default: \code{0.002})
m/z tolerance in Daltons for MS1 data.}

\item{mz_tol_da_ms2}{\code{numeric(1)} (default: \code{0.005})
m/z tolerance in Daltons for MS2 data.}

\item{mz_tol_ppm_ms1}{\code{numeric(1)} (default: \code{5})
m/z tolerance in parts per million (ppm) for MS1.}

\item{mz_tol_ppm_ms2}{\code{numeric(1)} (default: \code{10})
m/z tolerance in parts per million (ppm) for MS2.}

\item{mz_fun}{\code{function} (default: \code{base::mean})
Function to aggregate m/z values within each peak group.
Ignored if \code{mz_weighted = TRUE}.}

\item{int_fun}{\code{function} (default: \code{base::max})
Function to aggregate peak intensities within each peak group.}

\item{mz_weighted}{\code{logical(1)} (default: \code{TRUE})
If \code{TRUE}, uses intensity-weighted mean for m/z value aggregation.}

\item{time_domain}{\code{logical(1)} (default: \code{TRUE})
If \code{TRUE}, uses square root for m/z weighting (TRUE for TOF).}

\item{intensity_exponent}{\code{numeric(1)} (default: \code{3})
If \code{mz_weighted=TRUE}, the exponent of the intensity for m/z weighting.}
}
\value{
\code{logical(1)}
Returns \code{TRUE} if centroiding was successful, otherwise \code{FALSE}.
}
\description{
This function processes an mzML file to apply centroiding with detailed controls
for peak picking, smoothing, and noise estimation. It allows fine-tuning of MS1 and MS2
peak detection, optimizing spectral data analysis for various experimental needs.
}
\details{
The function processes both MS1 and MS2 data with user-defined smoothing, peak-picking, and noise estimation.
File path modifications are supported via \code{pattern} and \code{replacement}.
}
\author{
Johannes Rainer, Adriano Rutz
}
