% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{centroid_one_file}
\alias{centroid_one_file}
\title{Centroid an mzML file with configurable parameters}
\usage{
centroid_one_file(
  file,
  pattern,
  replacement,
  mz_tol_da = 0,
  mz_tol_ppm = 5,
  mz_fun = base::mean,
  int_fun = base::sum,
  mz_weighted = TRUE
)
}
\arguments{
\item{file}{\code{character(1)}
Path to the input mzML file.
The file must be accessible and in valid mzML format.}

\item{pattern}{\code{character(1)}
Regular expression pattern to match in the input file path, useful for modifying the output file path.}

\item{replacement}{\code{character(1)}
Replacement string for altering the output file path based on the \code{pattern}.}

\item{mz_tol_da}{\code{numeric(1)}
m/z tolerance in Dalton.

Default: \code{0}.}

\item{mz_tol_ppm}{\code{numeric(1)}
m/z tolerance in ppm.

Default: \code{5}.}

\item{mz_fun}{Function to aggregate m/z values for all mass peaks within each peak group into a single m/z value.
This parameter is ignored if weighted = TRUE (the default).

Default: \code{base::mean}.}

\item{int_fun}{Function to aggregate intensities for all peaks in each peak group into a single intensity value.

Default: \code{base::sum}.}

\item{mz_weighted}{\code{logical(1)}
Boolean whether m/z values of peaks within each peak group should be aggregated into a single m/z value using an intensity-weighted mean.

Default: \code{TRUE}.}
}
\value{
\code{logical(1)}
Returns \code{TRUE} if centroiding was successful; otherwise, returns \code{FALSE}.
}
\description{
This function processes an mzML file for centroiding, providing detailed controls for
peak picking, smoothing, and noise estimation. It allows fine-tuning of MS1 and MS2
peak detection, optimizing spectral data analysis for different experimental needs.
}
\details{
This function processes both MS1 and MS2 data, applying user-defined smoothing, peak-picking, and noise estimation methods.
The centroiding is tailored according to specific experimental requirements, with file path customization available via \code{pattern} and \code{replacement}.
}
\author{
Johannes Rainer, Adriano Rutz
}
