% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{centroid_one_file}
\alias{centroid_one_file}
\title{Centroid an mzML file with Configurable Peak Picking Parameters}
\usage{
centroid_one_file(
  file,
  pattern,
  replacement,
  fixed = TRUE,
  smooth_method = "SavitzkyGolay",
  smooth_window = 6L,
  refine_mz = "descendPeak",
  ms1_peak_snr = 1L,
  ms1_signal_percentage = 33,
  ms2_peak_snr = 1L,
  ms2_signal_percentage = 50,
  min_peaks = 1000
)
}
\arguments{
\item{file}{Path to the input mzML file}

\item{pattern}{Pattern to replace in file path}

\item{replacement}{Replacement string for output path}

\item{fixed}{Logical, whether pattern matching should be fixed}

\item{smooth_method}{Smoothing method for spectra}

\item{smooth_window}{Half window size for smoothing}

\item{refine_mz}{Refining method}

\item{ms1_peak_snr}{Signal-to-noise ratio for MS1 peak picking}

\item{ms1_signal_percentage}{Signal percentage for MS1 peak refinement}

\item{ms2_peak_snr}{Signal-to-noise ratio for MS2 peak picking}

\item{ms2_signal_percentage}{Signal percentage for MS2 peak refinement}

\item{min_peaks}{Minimum number of peaks to retain a spectrum}
}
\value{
Logical indicating success of centroiding
}
\description{
Centroid an mzML file with Configurable Peak Picking Parameters
}
\author{
Johannes Rainer
}
