% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{setup_logger}
\alias{setup_logger}
\title{Setup Logging for Centroiding Process}
\usage{
setup_logger(dir = Sys.getenv("HOME"), filename = "centroiding.log")
}
\arguments{
\item{dir}{character(1). Directory for saving the log file. Defaults to the output directory.}

\item{filename}{character(1). Log file name. Default: "centroiding.log".}
}
\description{
Setup Logging for Centroiding Process
}
\keyword{internal}
