% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_one_file.R
\name{try_centroid_one_file}
\alias{try_centroid_one_file}
\title{Wrapper for centroiding with additional error handling and flexible parameters}
\usage{
try_centroid_one_file(file, pattern, replacement, ...)
}
\arguments{
\item{file}{Path to the input mzML file}

\item{pattern}{Pattern to replace in file path}

\item{replacement}{Replacement string for output path}

\item{...}{Additional parameters to pass to centroid_one_file}
}
\value{
Logical indicating success of processing
}
\description{
Wrapper for centroiding with additional error handling and flexible parameters
}
\keyword{internal}
